Add .readthedocs.yml and update project metadata in pyproject.toml
A configuration file for Read the Docs has been created with .readthedocs.yml. Additionally, metadata in the project's pyproject.toml file, including the homepage, repository, and classifers, has been updated to improve project information and visibility. Extras dependencies for 'slack' and 'celery' were added to the poetry tool as well.