Refactor task queuing service to support conditional imports
The task queuing service has been refactored to support conditional imports of 'django_rq' and 'celery'. If 'django_rq' is not found, it will try to import 'celery' and if both are not found, it will raise an ImportError. The test cases have also been updated to accommodate this change.